requirements.txt
streamlit 
google-generativeai 
langchain-google-genai 
langchain-community 
python-dotenv 
pymysql
fastapi
uvicorn
pydantic
pydantic-settings
